ai-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow==11.0.0
numpy==2.1.2
python-multipart==0.0.12
httpx==0.27.2
onnxruntime==1.19.2

# Dependencies for dermatology_module
torch>=2.0.0
open_clip_torch>=2.20.0
